move tests into __tests__ directory
