+++ src/components/app-nav-menu/get-started/__tests__/SampleCard.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, RenderResult, fireEvent } from '@testing-library/react';
-import { SampleCard } from './SampleCard';
+import { SampleCard } from '../SampleCard';
 
 describe('Devfile Metadata Card', () => {
